changed version to uint16

+++ producer/src/accounts/accounts.go
@@ -18,7 +18,7 @@ Fills struct fields with parameters given
 Calls sign contract
 Returns contract
 */
-func MakeContract(version uint32, sender ecdsa.PrivateKey, recipient ecdsa.PublicKey, value uint64, nonce uint64) Contract {
+func MakeContract(version uint16, sender ecdsa.PrivateKey, recipient ecdsa.PublicKey, value uint64, nonce uint64) Contract {
 	return Contract{}
 }
